PA4: ready to go

+++ pa4/src/miniJava/CodeGenerator/RuntimeEntity.java
@@ -1,5 +1,5 @@
 package miniJava.CodeGenerator;
 
 public abstract class RuntimeEntity { 
-    public int size;
+    public short size;
 }
